+++ tests/integ/modin/frame/test_setitem.py
@@ -9,11 +9,13 @@ import numpy as np
 import pandas as native_pd
 import pytest
 from modin.pandas.utils import is_scalar
+from pytest import param
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
 from tests.integ.modin.utils import (
     assert_frame_equal,
     assert_snowpark_pandas_equals_to_pandas_without_dtypecheck,
+    create_test_dfs,
     eval_snowpark_pandas_result,
     try_cast_to_snowpark_pandas_series,
 )
@@ -46,14 +48,14 @@ SCALAR_LIKE_VALUES = [0, "xyz", None, 3.14]
         native_pd.Series(["b", "a"]),
     ],
 )
-def test_df_setitem_df_value(key):
+@pytest.mark.parametrize("dtype", ["int", "timedelta64[ns]"])
+def test_df_setitem_df_value(key, dtype):
     data = {"a": [1, 2, 3], "b": [4, 5, 6]}
-    snow_df = pd.DataFrame(data)
-    native_df = native_pd.DataFrame(data)
+    snow_df, native_df = create_test_dfs(data, dtype=dtype)
     val = (
-        native_pd.DataFrame({"a": [10, 20, 30]})
+        native_pd.DataFrame({"a": [10, 20, 30]}, dtype=dtype)
         if is_scalar(key)
-        else native_pd.DataFrame({"a": [10, 20, 30], "c": [40, 50, 60]})
+        else native_pd.DataFrame({"a": [10, 20, 30], "c": [40, 50, 60]}, dtype=dtype)
     )
 
     def setitem(df):
@@ -126,12 +128,14 @@ def test_df_setitem_df_value_dedup_columns(key, key_type):
         slice("-10", "100"),
     ],
 )
-def test_df_setitem_slice_key_df_value(key):
+@pytest.mark.parametrize("dtype", ["int", "timedelta64[ns]"])
+def test_df_setitem_slice_key_df_value(key, dtype):
     data = {"a": [1, 2, 3], "b": [4, 5, 6]}
     index = ["0", "1", "2"]
-    snow_df = pd.DataFrame(data, index=index)
-    native_df = native_pd.DataFrame(data, index=index)
-    val = native_pd.DataFrame({"a": [10, 20, 30], "c": [40, 50, 60]}, index=index)
+    snow_df, native_df = create_test_dfs(data, index=index, dtype=dtype)
+    val = native_pd.DataFrame(
+        {"a": [10, 20, 30], "c": [40, 50, 60]}, index=index, dtype=dtype
+    )
     val = val[key]
 
     def setitem(df):
@@ -141,9 +145,7 @@ def test_df_setitem_slice_key_df_value(key):
         else:
             df[key] = val
 
-    expected_join_count = 3 if isinstance(key.start, int) else 4
-
-    with SqlCounter(query_count=1, join_count=expected_join_count):
+    with SqlCounter(query_count=1, join_count=3):
         eval_snowpark_pandas_result(snow_df, native_df, setitem, inplace=True)
 
 
@@ -168,14 +170,16 @@ def test_df_setitem_slice_key_df_value(key):
         "A",
     ],
 )  # matching_item_columns_by_label is always True
-def test_df_setitem_df_single_value(key, val_index, val_columns):
+@pytest.mark.parametrize("dtype", ["int", "timedelta64[ns]"])
+def test_df_setitem_df_single_value(key, val_index, val_columns, dtype):
     native_df = native_pd.DataFrame(
         [[91, -2, 83, 74], [95, -6, 87, 78], [99, -10, 811, 712], [913, -14, 815, 716]],
         index=["x", "x", "z", "w"],
         columns=["A", "B", "C", "D"],
+        dtype=dtype,
     )
 
-    val = native_pd.DataFrame([100], columns=val_columns, index=val_index)
+    val = native_pd.DataFrame([100], columns=val_columns, index=val_index, dtype=dtype)
 
     def setitem(df):
         if isinstance(df, pd.DataFrame):
@@ -330,7 +334,14 @@ def test_df_setitem_self_df_set_aligned_row_key(native_df):
             index=[100, 101, 102]
         ),  # non-matching index will replace with NULLs
         native_pd.Series([]),
-        ["a", "c", "b"],  # replace with different type
+        param(["a", "c", "b"], id="string_type"),
+        param(
+            [pd.Timedelta(5), pd.Timedelta(6), pd.Timedelta(7)],
+            id="timedelta_type",
+            marks=pytest.mark.xfail(
+                strict=True, raises=NotImplementedError, reason="SNOW-1738952"
+            ),
+        ),
         native_pd.Series(["x", "y", "z"], index=[2, 0, 1]),
         native_pd.RangeIndex(3),
         native_pd.Index(
@@ -540,13 +551,26 @@ def test_df_setitem_full_columns(key, value):
 
 
 @sql_count_checker(query_count=1)
-def test_df_setitem_lambda_dataframe():
-    data = {"a": [1, 2, 3], "b": [4, 5, 6]}
+@pytest.mark.parametrize(
+    "data, comparison_value, set_value",
+    [
+        ({"a": [1, 2, 3], "b": [4, 5, 6]}, 2, 8),
+        (
+            {
+                "a": native_pd.to_timedelta([1, 2, 3]),
+                "b": native_pd.to_timedelta([4, 5, 6]),
+            },
+            pd.Timedelta(2),
+            pd.Timedelta(8),
+        ),
+    ],
+)
+def test_df_setitem_lambda_dataframe(data, comparison_value, set_value):
     snow_df = pd.DataFrame(data)
     native_df = native_pd.DataFrame(data)
 
     def masking_function(df):
-        df[lambda x: x < 2] = 8
+        df[lambda x: x < comparison_value] = set_value
 
     eval_snowpark_pandas_result(snow_df, native_df, masking_function, inplace=True)
 
@@ -1294,6 +1318,18 @@ def test_df_setitem_2d_array(indexer, item_type):
             setitem_helper,
             inplace=True,
         )
+
+
+@pytest.mark.xfail(strict=True, raises=NotImplementedError, reason="SNOW-1738952")
+def test_df_setitem_2d_array_timedelta_negative():
+    def setitem(df):
+        df[[1]] = np.array([[pd.Timedelta(3)]])
+
+    eval_snowpark_pandas_result(
+        *create_test_dfs(native_pd.DataFrame([[pd.Timedelta(1), pd.Timedelta(2)]])),
+        setitem,
+        inplace=True
+    )
 
 
 def test_df_setitem_2d_array_row_length_no_match():